shift files to specified directories
